+++ src/main/java/com/secuconnect/client/model/LoyaltyMerchantcardsLockStatus.java
@@ -4,19 +4,36 @@ import com.google.gson.annotations.SerializedName;
 import java.util.Objects;
 
 /**
- * LoyaltyMerchantcardsDTOLock
+ * LoyaltyMerchantcardsGetLock
  */
-public class LoyaltyMerchantcardsDTOLock {
+public class LoyaltyMerchantcardsLockStatus {
+  @SerializedName("lock_status")
+  protected String lockStatus = null;
+
   @SerializedName("reason")
   protected Integer reason = null;
 
   @SerializedName("note")
   protected String note = null;
 
-  @SerializedName("csc")
-  protected String csc = null;
+  public LoyaltyMerchantcardsLockStatus lockStatus(String lockStatus) {
+    this.lockStatus = lockStatus;
+    return this;
+  }
 
-  public LoyaltyMerchantcardsDTOLock reason(Integer reason) {
+   /**
+   * Lock status:  - &#x60;\&quot;locked\&quot;&#x60; - &#x60;\&quot;unlocked\&quot;&#x60;
+   * @return lockStatus
+  **/
+  public String getLockStatus() {
+    return lockStatus;
+  }
+
+  public void setLockStatus(String lockStatus) {
+    this.lockStatus = lockStatus;
+  }
+
+  public LoyaltyMerchantcardsLockStatus reason(Integer reason) {
     this.reason = reason;
     return this;
   }
@@ -33,7 +50,7 @@ public class LoyaltyMerchantcardsDTOLock {
     this.reason = reason;
   }
 
-  public LoyaltyMerchantcardsDTOLock note(String note) {
+  public LoyaltyMerchantcardsLockStatus note(String note) {
     this.note = note;
     return this;
   }
@@ -50,23 +67,6 @@ public class LoyaltyMerchantcardsDTOLock {
     this.note = note;
   }
 
-  public LoyaltyMerchantcardsDTOLock csc(String csc) {
-    this.csc = csc;
-    return this;
-  }
-
-   /**
-   * Card Security Code (optional in the most cases)
-   * @return csc
-  **/
-  public String getCsc() {
-    return csc;
-  }
-
-  public void setCsc(String csc) {
-    this.csc = csc;
-  }
-
   @Override
   public boolean equals(Object o) {
     if (this == o) {
@@ -75,24 +75,24 @@ public class LoyaltyMerchantcardsDTOLock {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    LoyaltyMerchantcardsDTOLock loyaltyMerchantcardsDTOLock = (LoyaltyMerchantcardsDTOLock) o;
-    return Objects.equals(this.reason, loyaltyMerchantcardsDTOLock.reason) &&
-        Objects.equals(this.note, loyaltyMerchantcardsDTOLock.note) &&
-        Objects.equals(this.csc, loyaltyMerchantcardsDTOLock.csc);
+    LoyaltyMerchantcardsLockStatus loyaltyMerchantcardsLockStatus = (LoyaltyMerchantcardsLockStatus) o;
+    return Objects.equals(this.lockStatus, loyaltyMerchantcardsLockStatus.lockStatus) &&
+        Objects.equals(this.reason, loyaltyMerchantcardsLockStatus.reason) &&
+        Objects.equals(this.note, loyaltyMerchantcardsLockStatus.note);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(reason, note, csc);
+    return Objects.hash(lockStatus, reason, note);
   }
 
   @Override
   public String toString() {
     StringBuilder sb = new StringBuilder();
-    sb.append("class LoyaltyMerchantcardsDTOLock {\n");
+    sb.append("class LoyaltyMerchantcardsLockStatus {\n");
+    sb.append("    lockStatus: ").append(toIndentedString(lockStatus)).append("\n");
     sb.append("    reason: ").append(toIndentedString(reason)).append("\n");
     sb.append("    note: ").append(toIndentedString(note)).append("\n");
-    sb.append("    csc: ").append(toIndentedString(csc)).append("\n");
     sb.append("}");
     return sb.toString();
   }